Refactor while loop if else conditions
