+++ src/main/java/ChadException.java
@@ -1,0 +1,5 @@
+public class ChadException extends Exception{
+    public ChadException(String message) {
+        super("☹ OOPS!!! " + message);
+    }
+}
